+++ src/commands/index.ts
@@ -1,0 +1,35 @@
+import {AdventOfCodeCommand} from '@/commands/misc/AdventOfCodeCommand';
+import {DeepFryCommand} from '@/commands/meme/DeepFryCommand';
+import {DrakeMemeCommand} from '@/commands/meme/DrakeMemeCommand';
+import {HelpCommand} from '@/commands/HelpCommand';
+import {JokeMemeCommand} from '@/commands/meme/JokeMemeCommand';
+import {KarmaCommand} from '@/commands/karma/KarmaCommand';
+import {KortebroekCommand} from '@/commands/fun/KortebroekCommand';
+import {MarieKondoCommand} from '@/commands/meme/MarieKondoCommand';
+import {MockifyCommand} from '@/commands/text/MockifyCommand';
+import {MotivationalQuoteCommand} from '@/commands/meme/MotivationalQuoteCommand';
+import {PingCommand} from '@/commands/PingCommand';
+import {PollCommand} from '@/commands/poll';
+import {SpooktoberCommand} from '@/commands/meme/SpooktoberCommand';
+import {StufiCommand} from '@/commands/fun/StufiCommand';
+import {UwuCommand} from '@/commands/text/UwuCommand';
+import {WhereMemeCommand} from '@/commands/meme/WhereMemeCommand';
+
+export default [
+    AdventOfCodeCommand,
+    DeepFryCommand,
+    DrakeMemeCommand,
+    HelpCommand,
+    JokeMemeCommand,
+    KarmaCommand,
+    KortebroekCommand,
+    MarieKondoCommand,
+    MockifyCommand,
+    MotivationalQuoteCommand,
+    PingCommand,
+    PollCommand,
+    SpooktoberCommand,
+    StufiCommand,
+    UwuCommand,
+    WhereMemeCommand,
+];