+++ src/commands/meme/SpooktoberCommand.ts
@@ -1,8 +1,10 @@
 import {EmbedBuilder} from 'discord.js';
 import Command from '@/commands/Command';
 import {getRandomGiphyGif} from '@/lib/giphy';
+import {trans} from '@/util/localization';
+import {getEnvString} from '@/util/environment';
 
-const {GIPHY_API_KEY} = process.env;
+const GIPHY_API_KEY = getEnvString('GIPHY_API_KEY', '');
 
 export const SpooktoberCommand = new Command({
     name: 'spooktober',
@@ -10,12 +12,14 @@ export const SpooktoberCommand = new Command({
     run: async (interaction) => {
         // 9 is October because starts at 0
         if (new Date().getMonth() !== 9) {
-            interaction.reply('Only available in October!!!');
+            interaction.reply(
+                trans('commands.spooktober.time_of_year_incorrect'),
+            );
             return;
         }
 
         const response = await getRandomGiphyGif({
-            api_key: GIPHY_API_KEY ?? '',
+            api_key: GIPHY_API_KEY,
             tag: 'halloween',
         });