+++ src/events/Event.ts
@@ -1,15 +1,19 @@
+import {BotContext} from '@/Bot';
 import type {ClientEvents} from 'discord.js';
 
 type Props<T extends keyof ClientEvents> = {
     name: T;
     enabled?: boolean;
-    run: (...args: ClientEvents[T]) => void | Promise<void>;
+    run: (ctx: BotContext, ...args: ClientEvents[T]) => void | Promise<void>;
 };
 
 export default class Event<T extends keyof ClientEvents> {
     public name: T;
     public enabled: boolean;
-    public run: (...args: ClientEvents[T]) => void | Promise<void>;
+    public run: (
+        ctx: BotContext,
+        ...args: ClientEvents[T]
+    ) => void | Promise<void>;
 
     constructor({name, enabled = true, run}: Props<T>) {
         this.name = name;